Refactor EventDetails page to match Members/MemberDetails architecture
- Extract EventDetails components into focused feature modules
- Create Header, Tabs, Overview, Marketing, Attendance, LoadingStates components
- Add useEventCrud hook for business logic separation
- Add useEventDetails hook to useEvents.ts with React Query
- Reduce EventDetails.tsx from 427 to 120 lines (~72% reduction)
- Follow consistent architecture patterns across all detail pages
