+++ administration/client/src/features/EventDetails/components/Attendance.tsx
@@ -1,0 +1,80 @@
+import { EventAttendance } from '@/types';
+
+interface AttendanceProps {
+  attendance: EventAttendance[];
+}
+
+export function Attendance({ attendance }: AttendanceProps) {
+  const formatDateTime = (dateTime: string) => {
+    return new Date(dateTime).toLocaleString();
+  };
+
+  if (attendance.length === 0) {
+    return (
+      <div>
+        <h3 className="text-lg font-medium mb-4">Event Attendance</h3>
+        <p className="text-gray-500 text-center py-8">No attendance records yet.</p>
+      </div>
+    );
+  }
+
+  return (
+    <div>
+      <h3 className="text-lg font-medium mb-4">Event Attendance</h3>
+      
+      <div className="overflow-x-auto">
+        <table className="min-w-full divide-y divide-gray-200">
+          <thead className="bg-gray-50">
+            <tr>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Member
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Ticket Type
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Registration
+              </th>
+              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                Check-in Status
+              </th>
+            </tr>
+          </thead>
+          <tbody className="bg-white divide-y divide-gray-200">
+            {attendance.map((record) => (
+              <tr key={record.id}>
+                <td className="px-6 py-4 whitespace-nowrap">
+                  <div className="text-sm font-medium text-gray-900">
+                    {record.member_id ? `Member ${record.member_id}` : 'Non-member'}
+                  </div>
+                  {record.eventbrite_attendee_id && (
+                    <div className="text-sm text-gray-500">
+                      Eventbrite: {record.eventbrite_attendee_id}
+                    </div>
+                  )}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
+                  {record.ticket_type || 'N/A'}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
+                  {record.registration_date ? formatDateTime(record.registration_date) : 'N/A'}
+                </td>
+                <td className="px-6 py-4 whitespace-nowrap">
+                  {record.check_in_method ? (
+                    <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800">
+                      Checked In ({record.check_in_method})
+                    </span>
+                  ) : (
+                    <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-yellow-100 text-yellow-800">
+                      Not Checked In
+                    </span>
+                  )}
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
+}
