+++ administration/client/src/features/EventDetails/components/LoadingStates.tsx
@@ -1,0 +1,34 @@
+interface LoadingStatesProps {
+  isLoading: boolean;
+  error: Error | null;
+}
+
+export function LoadingStates({ isLoading, error }: LoadingStatesProps) {
+  if (isLoading) {
+    return (
+      <div className="flex justify-center items-center h-64">
+        <div className="text-gray-500">Loading event details...</div>
+      </div>
+    );
+  }
+
+  if (error) {
+    return (
+      <div className="text-center py-12">
+        <div className="text-red-600 mb-4">Failed to load event</div>
+        <button
+          onClick={() => window.location.reload()}
+          className="bg-blue-600 text-white px-4 py-2 rounded-md hover:bg-blue-700"
+        >
+          Retry
+        </button>
+      </div>
+    );
+  }
+
+  return (
+    <div className="text-center py-12">
+      <h2 className="text-2xl font-semibold text-gray-900 mb-4">Event not found</h2>
+    </div>
+  );
+}
