+++ administration/client/src/features/EventDetails/components/Tabs.tsx
@@ -1,0 +1,44 @@
+import { ReactNode } from 'react';
+
+interface TabsProps {
+  activeTab: 'overview' | 'marketing' | 'volunteers' | 'attendance';
+  onTabChange: (tab: 'overview' | 'marketing' | 'volunteers' | 'attendance') => void;
+  volunteersCount: number;
+  attendanceCount: number;
+  children: ReactNode;
+}
+
+export function Tabs({ activeTab, onTabChange, volunteersCount, attendanceCount, children }: TabsProps) {
+  const tabs = [
+    { key: 'overview', label: 'Overview' },
+    { key: 'marketing', label: 'Marketing' },
+    { key: 'volunteers', label: `Volunteers (${volunteersCount})` },
+    { key: 'attendance', label: `Attendance (${attendanceCount})` }
+  ] as const;
+
+  return (
+    <div className="bg-white shadow rounded-lg">
+      <div className="border-b border-gray-200">
+        <nav className="flex space-x-8 px-6">
+          {tabs.map((tab) => (
+            <button
+              key={tab.key}
+              onClick={() => onTabChange(tab.key)}
+              className={`py-4 px-1 border-b-2 font-medium text-sm ${
+                activeTab === tab.key
+                  ? 'border-blue-500 text-blue-600'
+                  : 'border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300'
+              }`}
+            >
+              {tab.label}
+            </button>
+          ))}
+        </nav>
+      </div>
+
+      <div className="p-6">
+        {children}
+      </div>
+    </div>
+  );
+}
