+++ administration/client/src/features/EventDetails/components/index.ts
@@ -1,0 +1,6 @@
+export { Header } from './Header';
+export { Tabs } from './Tabs';
+export { Overview } from './Overview';
+export { Marketing } from './Marketing';
+export { Attendance } from './Attendance';
+export { LoadingStates } from './LoadingStates';
