+++ administration/client/src/features/Events/components/EventGrid.tsx
@@ -1,17 +1,11 @@
 import { CalendarIcon } from '@heroicons/react/24/outline';
 import { Event } from '../../../types';
-import { EventTableRow } from './EventTableRow';
 
 type EventFilter = 'upcoming' | 'past' | 'all';
 
 interface EventGridProps {
   events: Event[];
   isLoading: boolean;
-  onEdit: (event: Event) => void;
-  onDelete: (event: Event) => void;
-  onView: (event: Event) => void;
-  isEditPending: boolean;
-  isDeletePending: boolean;
   filter: EventFilter;
   onFilterChange: (filter: EventFilter) => void;
   children?: React.ReactNode;
@@ -20,11 +14,6 @@ interface EventGridProps {
 export function EventGrid({
   events,
   isLoading,
-  onEdit,
-  onDelete,
-  onView,
-  isEditPending,
-  isDeletePending,
   filter,
   onFilterChange,
   children,
@@ -33,31 +22,31 @@ export function EventGrid({
     <div className="bg-white shadow rounded-lg">
       <div className="p-6 border-b border-gray-200">
         <div className="flex items-center space-x-4">
-          <button 
+          <button
             onClick={() => onFilterChange('upcoming')}
             className={`px-4 py-2 rounded-md ${
-              filter === 'upcoming' 
-                ? 'bg-blue-100 text-blue-700' 
+              filter === 'upcoming'
+                ? 'bg-blue-100 text-blue-700'
                 : 'text-gray-600 hover:bg-gray-100'
             }`}
           >
             Upcoming
           </button>
-          <button 
+          <button
             onClick={() => onFilterChange('past')}
             className={`px-4 py-2 rounded-md ${
-              filter === 'past' 
-                ? 'bg-blue-100 text-blue-700' 
+              filter === 'past'
+                ? 'bg-blue-100 text-blue-700'
                 : 'text-gray-600 hover:bg-gray-100'
             }`}
           >
             Past
           </button>
-          <button 
+          <button
             onClick={() => onFilterChange('all')}
             className={`px-4 py-2 rounded-md ${
-              filter === 'all' 
-                ? 'bg-blue-100 text-blue-700' 
+              filter === 'all'
+                ? 'bg-blue-100 text-blue-700'
                 : 'text-gray-600 hover:bg-gray-100'
             }`}
           >
@@ -67,14 +56,14 @@ export function EventGrid({
       </div>
 
       {isLoading ? (
-        <div className="p-6 text-center text-gray-500">
-          Loading events...
-        </div>
+        <div className="p-6 text-center text-gray-500">Loading events...</div>
       ) : events.length === 0 ? (
         <div className="p-6">
           <div className="text-center text-gray-500">
             <CalendarIcon className="mx-auto h-12 w-12 text-gray-400 mb-4" />
-            <h3 className="text-lg font-medium text-gray-900 mb-2">No events scheduled</h3>
+            <h3 className="text-lg font-medium text-gray-900 mb-2">
+              No events scheduled
+            </h3>
             <p>Get started by creating your first event.</p>
           </div>
         </div>
@@ -101,22 +90,11 @@ export function EventGrid({
               </tr>
             </thead>
             <tbody className="bg-white divide-y divide-gray-200">
-              {events.map((event) => (
-                <EventTableRow
-                  key={event.id}
-                  event={event}
-                  onEdit={onEdit}
-                  onDelete={onDelete}
-                  onView={onView}
-                  isEditPending={isEditPending}
-                  isDeletePending={isDeletePending}
-                />
-              ))}
+              {children}
             </tbody>
           </table>
         </div>
       )}
-      {children}
     </div>
   );
 }
